pyproject.toml
[project]
name = "tcb-rca-framework"
version = "1.0.0"
description = "Multi-Agent Root Cause Analysis Framework with TCB-RCA Algorithm (MCP + A2A)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
keywords = ["rca", "root-cause-analysis", "microservices", "mcp", "multi-agent"]
dependencies = [
    "fastapi>=0.115",
    "httpx>=0.28",
    "pydantic>=2.7",
    "uvicorn>=0.30",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
eval = [
    "numpy>=1.26",
    "pandas>=2.1",
]
llm = [
    "openai>=1.50",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]

# =========================================================================
# uv-specific configuration
# =========================================================================
[tool.uv]
# 실험 재현을 위해 패키지 버전을 일관되게 관리
# (uv.lock은 gitignore에 포함되므로 pyproject.toml의 버전 범위가 진실의 원천)
package = false          # 라이브러리가 아닌 애플리케이션 프로젝트
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.uv.pip]
# uv pip install 기본 동작 설정
strict = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "mcp_servers", "common"]
